% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_bayesbox.R
\name{bayesbox}
\alias{bayesbox}
\title{Compute Bayes Box}
\usage{
bayesbox(hyps, priors, liks, round = 2L)
}
\arguments{
\item{hyps}{Vector of hypotheses (all types allowed), serves as ID}

\item{priors}{Vector of prior probabilites (numeric)}

\item{liks}{Vector of likelihoods (numeric)}

\item{round}{number of decimal places, NULL for not rounding (integer)}
}
\value{
data.frame with Bayes Box
}
\description{
Similar to Allen Downey's concept of a Bayes Box, such a table is computed by this function.
It can be used for simple Bayes models of discrete (non-continuous) variables.
}
\examples{

hyps <- seq(0,1, by = .1)
bayesbox(hyps = hyps, priors = 1, liks = dbinom(x = 5, size = 7, prob = hyps, round = 3))
}
