% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Cleans social media text}
\usage{
clean_text(x, ascify = FALSE)
}
\arguments{
\item{x}{character (not vectorized)}

\item{ascify}{should non-ASCII glyphs should be removed? (Defaults to FALSE)}
}
\value{
cleaned character
}
\description{
Removes typical artifacts in social media text such as "rt", "http",
but also spaces. No external packages are needed.
}
\details{
Expects UTF-8 as input format.
}
\examples{
clean_text("  Hello http @sesa rt 123  ")
}
