% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_count_lexicon.R
\name{count_lexicon}
\alias{count_lexicon}
\title{Count hits in a lexicon}
\usage{
count_lexicon(txt, lexicon)
}
\arguments{
\item{txt}{String in which the occurences of words from lexicon should be counted}

\item{lexicon}{vector of target words which should be counted in \code{text}}
}
\value{
string
}
\description{
Counts the number of times a string is detected in a lexicon,
according to a (regex) pattern
}
\examples{
my_txt <- c("Hello you! Too bad you are not here")
my_dict <- c("Bad", "good")
count_lexicon(my_txt, my_dict)
1
}
