% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_count_lexicon_df.R
\name{count_lexicon_df}
\alias{count_lexicon_df}
\title{Count hits in a lexicon}
\usage{
count_lexicon_df(data, lexicon, col, group)
}
\arguments{
\item{data}{tibble where text data (corpus) is stored}

\item{lexicon}{vector of target words which should be counted in \code{text}}

\item{col}{String in which the occurences of words from lexicon should be counted}

\item{group}{grouping variable}
}
\value{
Number of hits of text in dictionary
}
\description{
Counts the number of times a string is detected in a lexicon,
according to a (regex) pattern
This function is based on a inner join of the text and the dictionary.
}
\examples{
my_txt <- c("Hello you! Too bad you are not here")
my_dict <- c("Bad", "good")
count_lexicon(my_txt, my_dict)
1
}
