% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_count_lexicon_vec.R
\name{count_lexicon_vec}
\alias{count_lexicon_vec}
\title{Count hits in a lexicon}
\usage{
count_lexicon_vec(vec, lexicon_vec)
}
\arguments{
\item{vec}{String (vector) in which the occurences of words from lexicon should be counted}

\item{lexicon_vec}{vector of target words which should be counted in \code{text}}
}
\value{
Number of hits of text in dictionary
}
\description{
Counts the number of times a string is detected in a lexicon,
according to a (regex) pattern.
This function is based on a inner join of the text and the dictionary.
}
\examples{
my_txt <- c("Hello you! Too bad you are not here")
my_dict <- c("Bad", "good")
count_lexicon_vec(my_txt, my_dict)
1
}
