\name{cframe}
\alias{cframe}

\title{A sample cytoFrame object - German
  Cancer Research Center Heidelberg -}
\description{
 Archived \code{cytoFrame} object from a MAP kinase screen conducted at the
 German Cancer Research Center Heidelberg. In the fluorescence channel 3
 the expression of a YFP tag and in channel 7 the activation
 state of ERK2 was measured.
}
\usage{##cytoFrame object, see examples for details}
\format{cytoFrame object}
\source{
 German Cancer Research Center
 Heidelberg, Germany
}
\examples{

data(cytoFrame)

}
\keyword{datasets}
