\name{cset}
\alias{cset}

\title{A sample cytoSet object - German
  Cancer Research Center Heidelberg -}
\description{
 Archived \code{cytoSet} object from a MAP kinase screen conducted at the
 German Cancer Research Center Heidelberg. In the fluorescence channel 3
 the expression of a YFP tag and in channel 7 the activation
 state of ERK2 was measured. The set contains measurements from 5 wells
 of a 96 well plate
}
\usage{##cytoSet object, see examples for details}
\format{cytoSet object}
\source{
 German Cancer Research Center
 Heidelberg, Germany
}
\examples{

data(cytoSet)

}
\keyword{datasets}
