% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_download_osf.R
\name{download_osf}
\alias{download_osf}
\title{A function to download data from www.osf.io}
\usage{
download_osf(repo)
}
\arguments{
\item{repo}{short link to repo, eg. "meyhp" (chr)}
}
\value{
data as data.frame
}
\description{
A simple wrapper around download for not having to type a rather cumbersome download link
}
\details{
This is a convenience function to download data from a specified OSF repo. Internet access is needed.
}
\examples{

 download_osf("meyhp")
}
