% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_find_funs.R
\name{find_funs}
\alias{find_funs}
\title{A function to identify the installed package(s) for a given function name}
\usage{
find_funs(f)
}
\arguments{
\item{f}{function name (chr)}
}
\value{
\code{package_name} packages(s) which the function is part of (chr),
    \code{builtin_package}  whether the package comes with standard R (a 'builtin' package; lgl),
    \code{loaded} whether that package is currently loaded/attached (lgl)
}
\description{
\code{find_funs} identifies the package(s) in which a given function name \code{f} exists.
}
\details{
The code of this function was inspired by code from Ben Bolker
    at \url{https://stackoverflow.com/questions/10553755/name-of-a-package-for-a-given-function-in-r}
}
\examples{

 find_funs("qplot")
 find_funs("sdfjkl")


}
