% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_get_dag_size.R
\name{get_dag_size}
\alias{get_dag_size}
\title{Get the size of a DAG}
\usage{
get_dag_size(dag)
}
\arguments{
\item{dag, }{a DAG object from `dagitty`}
}
\value{
dag size (integer)
}
\description{
Returns the number of variables in a directed acyclic graph (DAG)
}
\examples{

get_dag_size(my_dag)

}
