% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_install_prada_pckgs.R
\name{install_prada_pckgs}
\alias{install_prada_pckgs}
\title{Get packages necessary for all analyses in package \code{prada}}
\usage{
install_prada_pckgs()
}
\description{
This function checks if you have all the packages necessary for the analyses of this pacakge. If you don't, they will be installed.
}
\examples{

 install_prada_pckgs()

}
