% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_is_in_tolerance.R
\name{is_in_tolerance}
\alias{is_in_tolerance}
\title{Is value within tolerance?}
\usage{
is_in_tolerance(asis, tobe, tol_rel, tol_abs)
}
\arguments{
\item{asis}{as-is value (num), the observed value}

\item{tobe}{to-be value (num), the target value}

\item{tol_rel}{relative tolerance (as a proportion, num)}

\item{tol_abs}{absolute tolerance (num)}
}
\value{
in_tolerenace, lgl
}
\description{
Computes if a value is within a tolerance region.
Both an absolute tolerance and a relative tolerance are considered.
If at least one of both criteria are met, the fuction returns TRUE,
otherwise FALSE.
}
\examples{
is_in_tolerance(asis = 39.5, tobe = 42, tol_rel = .05, tol_abs = 2.5)
}
