% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_most.R
\name{most}
\alias{most}
\title{A function to extract the mode}
\usage{
most(x)
}
\arguments{
\item{x}{vector (numeric or integer)}
}
\value{
mode most frequent value
}
\description{
\code{most} extracts the most frequent value (mfv; mode) of a numeric vector.
}
\details{
This is a convenience function to get the mode of a (numeric) vector.
}
\examples{

 most(c(1,1,2))
 most(mtcars$cyl)
}
