% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_na_count.R
\name{na_count}
\alias{na_count}
\title{Count all missing values in a data frame}
\usage{
na_count(data)
}
\arguments{
\item{data}{data frame or tibble}
}
\value{
count of all NAs
}
\description{
\code{na_count} counts the missing values in all columns of a data frame.
}
