\name{Internal functions}
\alias{gatePoints}
%\alias{}
\title{Internal functions called by other gating functions}
\description{
  These functions are for internal use by other gating functions and not
  to be called by the user.
}
%\usage{}
%\arguments{}
%\details{}
%\value{}
%\references{ ~put references to the literature/web site here ~ }
\author{Joern Toedling \email{toedling@ebi.ac.uk}}
%\seealso{}
%\examples{}
\keyword{internal}% at least one, from doc/KEYWORDS
