% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_pdf2png.R
\name{pdf2png}
\alias{pdf2png}
\title{Convert pdf file to png file, with white space cropped}
\usage{
pdf2png(pdf_path)
}
\arguments{
\item{pdf_path}{path to folder where to find pdf images}
}
\description{
converts all pdfs in the path to png
crops whites space
tries to maintain high quality
calls image magick convert
}
