% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prada.R
\docType{package}
\name{prada}
\alias{prada}
\alias{prada-package}
\title{prada: A package accompanying the course "practical data analysis" (hence pra-da) and the book "stastisics_21"}
\description{
The package contains some conveniecne functions and the data used in the aforementioned course and book
}
