% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_prada_pckgs.R
\docType{data}
\name{prada_pckgs}
\alias{prada_pckgs}
\title{Vector of all packages needed for "Praktische Datenanalyse" and "Statistik_21"}
\format{
An object of class \code{character} of length 40.
}
\usage{
prada_pckgs
}
\description{
Vector of all packages needed for "Praktische Datenanalyse" and "Statistik_21"
}
\keyword{datasets}
