% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_prop_fav.R
\name{prop_fav}
\alias{prop_fav}
\title{A function to compute the proportion of 'favorable' pairwise comparisons}
\usage{
prop_fav(df, value, group, g1, g2)
}
\arguments{
\item{df}{data frame}

\item{value}{are (unquoted) column with values of the two vectors}

\item{group}{bare (unquoted) column with grouping information}

\item{g1}{name of group 1 (quoted if not numeric)}

\item{g2}{name of group 2 (quoted if not numeric)}
}
\value{
proportion of x>y
}
\description{
This function takes two vectors (x,y; ie., group 1 and 2) of values and computes all pairwise comparisons (>, ie., "greater than"). The proportion of x>y is returned. Columns must be quoted.
}
\examples{
prop_fav(df = mtcars, value = hp, group = cyl, g1 = 6, g2 = 4)
data(extra, package = "prada")
prop_fav(df = extra, value = extra_mean, group = sex, g1 = "Frau", g2 = "Mann")
}
