% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAME_sample_size_for_r2.R
\name{sample_size_for_r2}
\alias{sample_size_for_r2}
\title{Sample size for a given R2 width}
\usage{
sample_size_for_r2(u, r2, target_width = 0.1)
}
\arguments{
\item{u}{number of predictors}

\item{r2}{observed or expected R squared}

\item{target_width}{width of the confidence interval (dafaults to 0.1)}
}
\value{
data frame with upper/lower limit of confidence interval
}
\description{
Computes sample size needed for a given width of a R squared confidence intervall.
}
\details{
In the contect of simple linear models, compute the minimal sample size given the parameters.
}
\examples{

u <- 5
r2 <- .3
sample_size_for_r2(u = u, r2 = r2)
}
