\name{statWell}
\alias{statWell}
\title{Calculate a statistic for the data from one well}
\description{Calculate a statistic for the data from one well}
\usage{statWell(x, pars, outdir, plotfile)}
\arguments{
  \item{x}{data.frame with columns \code{brdu}, \code{trsf},
    \code{dapi}, \code{Field}}
  \item{pars}{list containing parameters for the calculation of the
    statistic}
  \item{plotfile}{Character. If not missing, a suitable plot is
    generated and saved as a PNG file under that name}
  \item{outdir}{Character. If not missing, the directory where plot file
    saved to}  
}

\details{}

\value{?}
\author{Wolfgang Huber \url{http://www.dkfz.de/abt0840/whuber}}
\keyword{manip}

